feat :`useStoreConnectionPlatforms` 훅 구현
