+++ frontend/src/pages/ConnectionPage/ConnectionPage.tsx
@@ -1,37 +1,12 @@
 import Spinner from 'components/@common/Spinner/Spinner';
-import { ConnectionPlatforms, getConnectionPlatformRedirectURL } from 'constants/components/myPage';
-import { usePageNavigate } from 'hooks/usePageNavigate';
 import { useEffect } from 'react';
-import { useLocation, useSearchParams } from 'react-router-dom';
 import { styled } from 'styled-components';
 import { useStoreConnectionPlatforms } from './useStoreConnectionPlatforms';
 
 const ConnectionPage = () => {
-  const { goMyPage } = usePageNavigate();
-  const location = useLocation();
-  const [searchParams] = useSearchParams();
-  const platform = location.pathname.split('/').pop();
-  const mutate = useStoreConnectionPlatforms(platform);
+  const storeInfo = useStoreConnectionPlatforms();
 
-  useEffect(() => {
-    const code = searchParams.get('code');
-
-    const isConnectionPlatforms = (
-      platform: string | undefined,
-    ): platform is ConnectionPlatforms => {
-      return platform ? platform in ConnectionPlatforms : false;
-    };
-
-    if (!code || !mutate || !isConnectionPlatforms(platform)) {
-      goMyPage();
-      return;
-    }
-
-    mutate({
-      code,
-      redirect_uri: getConnectionPlatformRedirectURL(platform),
-    });
-  }, []);
+  useEffect(storeInfo, []);
 
   return (
     <S.Section>
